improve dp doc
