+++ dev/CustomPlaygrounds/dp-validation.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
 	import { ref, watch, computed } from 'vue'
-	import DatePicker from '@/components/DatePicker/DatePicker.vue'
+	import DatePicker from '@/components/DatePicker/DatePicker/DatePicker.vue'
 	import { useDateFormat } from '@/composables/date/useDateFormat'
 
 	const { parseDate } = useDateFormat()
