--- src/components/DatePicker/DatePicker.mdx
@@ -1,222 +1,0 @@
-import {Meta, Canvas, Controls, Source} from '@storybook/blocks';
-import * as DatePickerStories from "./DatePicker.stories.ts";
-
-<Meta of={DatePickerStories}/>
-
-# DatePicker
-
-Le composant `DatePicker` est disponible en deux variantes :
-- La version avec un calendrier interactif pour sélectionner une date ou un range,
-- La version (avec la prop `no-calendar`) qui propose uniquement un champ de saisie textuelle pour entrer la date manuellement et permettre aux robots de la renseigner automatiquement.
-
-Nous ferons évoluer le composant pour qu'il propose les 2 en même temps.
-
-<Canvas story={{height: '550px'}} of={DatePickerStories.Default}/>
-
-# API
-
-<Controls of={DatePickerStories.Default}/>
-
-# Props
-
-## Valeur et contrôle
-
-### modelValue
-- **Type** : `string | string[]`
-- **Description** : La valeur du champ. Pour une plage de dates, utiliser un tableau de deux dates. La valeur doit être fournie dans le même format que celui spécifié dans la prop `format`.
-- **Défaut** : `undefined`
-
-### placeholder
-- **Type** : `string`
-- **Description** : Texte indicatif affiché quand le champ est vide
-- **Défaut** : `'Sélectionner une date'`
-
-### format
-- **Type** : `string`
-- **Description** : Format d'affichage et d'entrée de la date. Supporte les formats suivants :
-- `DD` : jour sur 2 chiffres
-- `MM` : mois sur 2 chiffres
-- `YYYY` : année sur 4 chiffres
-- `YY` : année sur 2 chiffres (assumé comme 20XX)
-Les séparateurs peuvent être `/`, `-` ou `.`
-- **Défaut** : `'DD/MM/YYYY'`
-
-### dateFormatReturn
-- **Type** : `string`
-- **Description** : Format de la date pour la valeur de retour. Utilise les mêmes tokens que la prop `format`. Si non spécifié, utilise le même format que `format`.
-- **Défaut** : `''`
-
-<div
-    style={{
-        border: '2px solid #FF0000',
-        padding: '1rem',
-        borderRadius: '4px',
-        marginBottom: '1rem',
-    }}
->
-    <strong>Note importante sur les formats</strong>
-    <p>
-        Le composant utilise deux formats distincts :
-    </p>
-    <ol>
-        <li>
-            <strong>Format d'entrée/affichage</strong><br/>
-            Défini par la prop <code>format</code>, il détermine à la fois :
-            <ul>
-                <li>Le format dans lequel la date doit être fournie au composant</li>
-                <li>Comment la date est affichée visuellement dans le champ</li>
-            </ul>
-        </li>
-        <li>
-            <strong>Format de retour</strong><br/>
-            Défini par la prop <code>dateFormatReturn</code>, il détermine le format de la date émise
-        </li>
-    </ol>
-    <p>
-        Par exemple, avec <code>format="DD-MM-YY"</code> et <code>dateFormatReturn="YYYY/MM/DD"</code> :
-    </p>
-    <ul>
-        <li>Vous devez fournir la date au format : <code>24-12-25</code></li>
-        <li>Elle sera affichée comme : <code>24-12-25</code></li>
-        <li>Elle sera retournée comme : <code>2025/12/24</code></li>
-    </ul>
-</div>
-
-
-
-### required
-- **Type** : `boolean`
-- **Description** : Indique si le champ est obligatoire
-- **Défaut** : `false`
-
-### disabled
-- **Type** : `boolean`
-- **Description** : Désactive le champ
-- **Défaut** : `false`
-
-## Configuration du calendrier
-
-### isBirthDate
-- **Type** : `boolean`
-- **Description** : Active le mode date de naissance (commence par la sélection de l'année)
-- **Défaut** : `false`
-
-### isOutlined
-- **Type** : `boolean`
-- **Description** : Active le style de champ de saisie "outlined" (bordure) plutôt que "underlined" (fond)
-- **Défaut** : `true`
-
-### showWeekNumber
-- **Type** : `boolean`
-- **Description** : Affiche les numéros de semaine dans le calendrier
-- **Défaut** : `false`
-
-### displayRange
-- **Type** : `boolean`
-- **Description** : Active la sélection d'une plage de dates
-- **Défaut** : `false`
-
-### noCalendar
-- **Type** : `boolean`
-- **Description** : Désactive l\'affichage du calendrier (saisie manuelle uniquement), elle permet les copier coller et le passage de robots.
-- **Défaut** : `false`
-
-## Utilisation de noCalendar
-
-L'option `noCalendar` est particulièrement utile dans les cas suivants :
-
-### Cas d'utilisation
-- **Saisie rapide** : Permet aux utilisateurs de saisir directement une date sans avoir à naviguer dans un calendrier
-- **Copier-coller** : Facilite le copier-coller de dates depuis d'autres sources
-- **Automatisation** : Compatible avec les tests automatisés et les robots qui doivent interagir avec le champ de date
-- **Accessibilité** : Offre une alternative pour les utilisateurs qui préfèrent ou doivent utiliser le clavier
-
-### Bonnes pratiques
-- Toujours spécifier le format de date attendu via la prop `format` pour guider l'utilisateur
-- Utiliser la prop `placeholder` pour indiquer le format de saisie attendu
-- Combiner avec la prop `required` si le champ est obligatoire
-- Penser à ajouter des règles de validation personnalisées si nécessaire
-
-## Apparence
-
-### displayIcon
-- **Type** : `boolean`
-- **Description** : Affiche l'icône de calendrier
-- **Défaut** : `false`
-
-### displayAppendIcon
-- **Type** : `boolean`
-- **Description** : Affiche l'icône à la fin du champ plutôt qu'au début
-- **Défaut** : `false`
-
-### noIcon
-- **Type** : `boolean`
-- **Description** : Masque toutes les icônes
-- **Défaut** : `false`
-
-### Exemple
-```vue
-<DatePicker
-  v-model="date"
-  no-calendar
-  format="DD/MM/YYYY"
-  placeholder="JJ/MM/AAAA"
-  required
-/>
-```
-
-## Validation
-
-### customRules
-- **Type** : `{ type: string, options: RuleOptions }[]`
-- **Description** : Règles de validation personnalisées
-- **Défaut** : `[]`
-
-### customWarningRules
-- **Type** : `{ type: string, options: RuleOptions }[]`
-- **Description** : Règles d'avertissement personnalisées
-- **Défaut** : `[]`
-
-# Exemple d'utilisation
-
-<Source
-    dark code={`
-<script setup lang="ts">
-	import { ref } from 'vue'
-	import { DatePicker } from '@cnamts/synapse'
-
-	const date = ref('')
-</script>
-
-<template>
-	<DatePicker
-		v-model="date"
-		placeholder="Sélectionner une date"
-		format="DD/MM/YYYY"
-		required
-	/>
-</template>
-`}
-/>
-
-## Exemple avec plage de dates
-
-<Source
-    dark code={`
-<script setup lang="ts">
-	import { ref } from 'vue'
-	import { DatePicker } from '@cnamts/synapse'
-
-	const dateRange = ref(['', ''])
-</script>
-
-<template>
-	<DatePicker
-		v-model="dateRange"
-		placeholder="Sélectionner une période"
-		format="DD/MM/YYYY"
-		display-range
-	/>
-</template>
-`}
-/>
