+++ src/components/DatePicker/DatePicker/DatePicker.stories.ts
@@ -4,7 +4,7 @@ import { ref, watch, computed } from 'vue'
 import { useDateFormat } from '@/composables/date/useDateFormatDayjs'
 
 const meta = {
-	title: 'Composants/Formulaires/DatePicker',
+	title: 'Composants/Formulaires/DatePicker/DatePicker',
 	component: DatePicker,
 	decorators: [
 		() => ({
@@ -797,405 +797,6 @@ export const WithDateFormatReturn: Story = {
                 />
               </div>
             `,
-		}
-	},
-}
-
-export const NoCalendarEuropeanFormat: Story = {
-	parameters: {
-		sourceCode: [
-			{
-				name: 'Template',
-				code: `
-				<template>
-					<DatePicker
-						v-model="date"
-						format="DD/MM/YYYY"
-						date-format-return="YYYY/MM/DD"
-						placeholder="JJ/MM/AAAA"
-						required
-						no-icon
-						no-calendar
-					/>
-				</template>
-				`,
-			},
-		],
-	},
-	args: {
-		noCalendar: true,
-		format: 'DD/MM/YYYY',
-		dateFormatReturn: 'YYYY/MM/DD',
-		placeholder: 'JJ/MM/AAAA',
-		required: true,
-		noIcon: true,
-	},
-	render(args) {
-		const date = ref<string | null>(null)
-		return {
-			components: { DatePicker },
-			setup() {
-				return { args, date }
-			},
-			template: `
-				<div style="padding: 20px;">
-					<h4 class="mb-4">Format européen avec règles de base (format de date valide) :</h4>
-					<DatePicker
-						v-model="date"
-						v-bind="args"
-					/>
-					<div style="margin-top: 10px; font-family: monospace; color: #666;">
-						Valeur (dateFormatReturn: 'YYYY/MM/DD') : {{ date }}
-					</div>
-				</div>
-			`,
-		}
-	},
-}
-
-export const NoCalendarCustomRules: Story = {
-	parameters: {
-		sourceCode: [
-			{
-				name: 'Template',
-				code: `
-				<template>
-					<DatePicker
-						v-model="date"
-						date-format-return="DD/MM/YYYY"
-						format="DD/MM/YYYY"
-						placeholder="DD/MM/YYYY"
-						required
-						no-calendar
-						:custom-rules="[{
-							type: 'custom',
-							options: {
-								validate: value => !value || !value.includes('2024'),
-								message: 'Les dates en 2024 ne sont pas autorisées',
-								successMessage: 'Les dates hors 2024 sont autorisées',
-								fieldIdentifier: 'date'
-							}
-						}]"
-					/>
-				</template>
-				`,
-			},
-		],
-	},
-	args: {
-		noCalendar: true,
-		format: 'DD/MM/YYYY',
-		dateFormatReturn: 'DD/MM/YYYY',
-		placeholder: 'DD/MM/YYYY',
-		required: true,
-		customRules: [{
-			type: 'custom',
-			options: {
-				validate: (value: string) => !value || !value.includes('2024'),
-				message: 'Les dates en 2024 ne sont pas autorisées',
-				successMessage: 'Les dates hors 2024 sont autorisées',
-				fieldIdentifier: 'date',
-			},
-		}],
-	},
-	render(args) {
-		const date = ref<string | null>('21/12/2024')
-		return {
-			components: { DatePicker },
-			setup() {
-				return { args, date }
-			},
-			template: `
-				<div style="padding: 20px;">
-					<h4 class="mb-0">Format avec règles personnalisées :</h4>
-					<p class="mb-4">Les dates en 2024 ne sont pas autorisées</p>
-					<DatePicker
-						v-model="date"
-						v-bind="args"
-					/>
-					<div style="margin-top: 10px; font-family: monospace; color: #666;">
-						Valeur : {{ date }}
-					</div>
-				</div>
-			`,
-		}
-	},
-}
-
-export const NoCalendarWarningRules: Story = {
-	parameters: {
-		sourceCode: [
-			{
-				name: 'Template',
-				code: `
-				<template>
-					<DatePicker
-						v-model="date"
-						format="DD/MM/YYYY"
-						placeholder="JJ/MM/AAAA"
-						no-calendar
-						:custom-warning-rules="[{
-							type: 'custom',
-							options: {
-								validate: value => !value || !value.includes('2025'),
-								warningMessage: 'Les dates en 2025 ne sont pas autorisées',
-								successMessage: 'Date hors 2025',
-								fieldIdentifier: 'date',
-								isWarning: true
-							}
-						}]"
-					/>
-				</template>
-				`,
-			},
-		],
-	},
-	args: {
-		noCalendar: true,
-		format: 'DD/MM/YYYY',
-		placeholder: 'JJ/MM/AAAA',
-		customWarningRules: [{
-			type: 'custom',
-			options: {
-				validate: (value: string) => !value || !value.includes('2025'),
-				warningMessage: 'Les dates en 2025 ne sont pas autorisées',
-				successMessage: 'Date hors 2025',
-				fieldIdentifier: 'date',
-				isWarning: true,
-			},
-		}],
-	},
-	render(args) {
-		const date = ref<string | null>('20/12/2025')
-		return {
-			components: { DatePicker },
-			setup() {
-				return { args, date }
-			},
-			template: `
-				<div style="padding: 20px;">
-					<h4 class="mb-0">Format avec règles d'avertissement :</h4>
-					<p class="mb-4">Les dates en 2025 ne sont pas autorisées</p>
-					<DatePicker
-						v-model="date"
-						v-bind="args"
-					/>
-					<div style="margin-top: 10px; font-family: monospace; color: #666;">
-						Valeur : {{ date }}
-					</div>
-				</div>
-			`,
-		}
-	},
-}
-
-export const WithErrorDisabled: Story = {
-	parameters: {
-		sourceCode: [
-			{
-				name: 'Template',
-				code: `
-				<template>
-					<DatePicker
-						v-model="date"
-						placeholder="Sélectionner une date"
-						format="DD/MM/YYYY"
-						:required="true"
-						:disableErrorHandling="true"
-					/>
-				</template>
-				`,
-			},
-			{
-				name: 'Script',
-				code: `
-				<script setup lang="ts">
-					import { ref } from 'vue'
-					import { DatePicker } from '@cnamts/synapse'
-					
-					const date = ref('')
-				</script>
-				`,
-			},
-		],
-	},
-	args: {
-		placeholder: 'Date requise sans affichage d\'erreur',
-		format: 'DD/MM/YYYY',
-		isBirthDate: false,
-		showWeekNumber: false,
-		required: true,
-		displayRange: false,
-		displayIcon: true,
-		displayAppendIcon: false,
-		disabled: false,
-		noIcon: false,
-		noCalendar: false,
-		disableErrorHandling: true,
-		modelValue: '',
-	},
-	render: (args) => {
-		return {
-			components: { DatePicker: DatePicker },
-			setup() {
-				const value = ref('')
-				return { args, value }
-			},
-			template: `
-              <div class="d-flex flex-column pa-4">
-                <div class="mb-5">
-                  <p class="mb-3">Ce champ est requis mais n'affiche pas de message d'erreur grâce à <code>disableErrorHandling</code>:</p>
-                  <DatePicker v-bind="args" v-model="value"/>
-                </div>
-                <div>
-                  <p class="mb-3">Comparaison avec un champ requis standard:</p>
-                  <DatePicker
-                    placeholder="Date requise avec erreur"
-                    format="DD/MM/YYYY"
-                    :required="true"
-                    v-model="value"
-                  />
-                </div>
-              </div>
-            `,
-		}
-	},
-}
-
-export const NoCalendarWithAppendIcon: Story = {
-	parameters: {
-		sourceCode: [
-			{
-				name: 'Template',
-				code: `
-				<template>
-					<DatePicker
-						v-model="date"
-						format="DD/MM/YYYY"
-						placeholder="JJ/MM/AAAA"
-						no-calendar
-						display-append-icon
-					/>
-				</template>
-				`,
-			},
-		],
-	},
-	args: {
-		noCalendar: true,
-		format: 'DD/MM/YYYY',
-		placeholder: 'JJ/MM/AAAA',
-		displayAppendIcon: true,
-	},
-	render(args) {
-		const date = ref<string | null>(null)
-		return {
-			components: { DatePicker },
-			setup() {
-				return { args, date }
-			},
-			template: `
-				<div style="padding: 20px;">
-					<h4 class="mb-4">Format avec icône en préfixe</h4>
-					<DatePicker
-						v-model="date"
-						v-bind="args"
-					/>
-					<div style="margin-top: 10px; font-family: monospace; color: #666;">
-						Valeur : {{ date }}
-					</div>
-				</div>
-			`,
-		}
-	},
-}
-
-export const NoCalendarWithErrorDisabled: Story = {
-	parameters: {
-		sourceCode: [
-			{
-				name: 'Template',
-				code: `
-				<template>
-					<div class="d-flex">
-						<div class="mr-4" style="width: 300px;">
-							<p class="mb-3">Avec <code>disableErrorHandling</code>:</p>
-							<DatePicker
-								v-model="date1"
-								format="DD/MM/YYYY"
-								placeholder="Date requise sans erreur"
-								required
-								no-icon
-								no-calendar
-								:disableErrorHandling="true"
-							/>
-						</div>
-						<div style="width: 300px;">
-							<p class="mb-3">Sans <code>disableErrorHandling</code>:</p>
-							<DatePicker
-								v-model="date2"
-								format="DD/MM/YYYY"
-								placeholder="Date requise avec erreur"
-								required
-								no-icon
-								no-calendar
-							/>
-						</div>
-					</div>
-				</template>
-				`,
-			},
-		],
-	},
-	args: {
-		noCalendar: true,
-		format: 'DD/MM/YYYY',
-		dateFormatReturn: 'YYYY/MM/DD',
-		placeholder: 'Date requise sans erreur',
-		required: true,
-		noIcon: true,
-		disableErrorHandling: true,
-	},
-	render(args) {
-		const date1 = ref<string | null>(null)
-		const date2 = ref<string | null>(null)
-		return {
-			components: { DatePicker },
-			setup() {
-				return { args, date1, date2 }
-			},
-			template: `
-				<div style="padding: 20px;">
-					<h4 class="mb-4">DateTextInput avec désactivation des erreurs</h4>
-					<div class="d-flex mb-4">
-						<div class="mr-4" style="width: 300px;">
-							<p class="mb-3">Avec <code>disableErrorHandling</code>:</p>
-							<DatePicker
-								v-model="date1"
-								v-bind="args"
-							/>
-							<div style="margin-top: 10px; font-family: monospace; color: #666;">
-								Valeur : {{ date1 }}
-							</div>
-						</div>
-						
-						<div style="width: 300px;">
-							<p class="mb-3">Sans <code>disableErrorHandling</code>:</p>
-							<DatePicker
-								v-model="date2"
-								format="DD/MM/YYYY"
-								placeholder="Date requise avec erreur"
-								required
-								no-icon
-								no-calendar
-							/>
-							<div style="margin-top: 10px; font-family: monospace; color: #666;">
-								Valeur : {{ date2 }}
-							</div>
-						</div>
-					</div>
-				</div>
-			`,
 		}
 	},
 }
