+++ src/components/DatePicker/docExamples/architecture.md
